Fix docs to have a valid doc build (#70)
Run `make docs` didn't work with the previous structure. I mainly fixed titles etc. to be a valid asciidoc format for our doc build. It is not the intention to fix the doc structure in this PR but being able to start with the doc build integration which requires a green doc build to start with.