Improve logging for production (#103)
In production I expect a log message on the error level that I have to do something. For our 400 response no changes are needed on the server as the problem is potentially on the agent side.

Now only if 500 happens an error is logged, all other are logged on debug level if enabled. Otherwise the production log could be flooded with error messages because of invalid requests or tokens which can be pretty normal if the server has a public API.