Fail setup dashboard sooner (#1999) (#2007)
* don't bundle dashboards

* help with setup.dashboards deprecation

Dashboard loading and the corresponding setup.dashboards configuration
have been deprecated since 6.4.  Since libbeat still supports that
option, apm-server will refuse to start when enabled, as libbeat will
fail to load dashboards that are no longer bundled.  To enable a
smoother transition to 7.0, apm-server will only log a warning when the
option is set, but false, which should be rare.