Flush aggregator on shutdown (#3971)
* publish: publish enqueued events on shutdown

Change the publisher to wait for enqueued events to
be published when Stop is called, up to the configured
ShutdownTimeout.

* aggregation/txmetrics: graceful shutdown

Remove the context param from Aggregator.Run, and
instead add a Stop method which accepts a context.
When Stop is called it directs Run to exit after
performing one final publication. Stop waits for
Run to exit, or for the context to be cancelled,
whichever comes first.

We call the Stop method with a timeout based on
the `apm-server.shutdown_timeout` configuration.

* tests/system: add test for flushing aggregations

* publish: add context to publisher.Stop

publisher.Stop now accepts a context, which is used
to interrupt waiting for published events to be
acknowledged. This is implemented by introducing a
custom beat.ACKer.

* Move publisher ShutdownTimeout handling to beater

* Update changelog