systemtest: compensate for Fleet API bug (#4693)
* systemtest: compensate for Fleet API bug

https://github.com/elastic/kibana/issues/90544 is
causing the system test to respond with 404 Not Found,
despite the policy existing and being successfully
deleted.

Update the test to temporarily expect a 404 response to
avoid failing all our builds. Once the Kibana bug is
resolved this will start failing, and we can again check
that no error is returned.