Remove testing/environments (#5035)
* Remove testing/environments

* docker-compose.yml: define services locally

We only use the snapshot services, and we customise
the services heavily; let's just define them in our
own docker-compose.yml and drop the dependency on
testing/environments.

* Make target to check docker images are up to date

`make check-docker-compose` checks if the images in
docker-compose.yml are up to date with the most
recent snapshot version for the current branch.

* docker-compose.yml: wait for yellow status

Change the Elasticsearch healthcheck to wait
for the status to become yellow, rather than
green. Yellow means primaries are assigned,
but one or more replicas are not. We don't
care about replicas in system tests.