sampling/pubsub: minor fixes (#5568)
* sampling/pubsub: close BulkIndexer

Revise the pubsub API so that the PublishSampledTraceIDs
method is long-lived and accepts a channel of trace IDs;
this brings it closer to the API of SubscribeSampledTraceIDs,
and ensures we properly close the BulkIndexer used by the
publisher when it exits.

* sampling/pubsub: fix double WriteHeader in test

Write HTTP response from the handler goroutine,
rather than in another goroutine which waits for
the request. This avoids a race in the existing
test code, where the client receives the response
and closes the connection, causing the server's
handler context to be cancelled before the handler
returns.