sourcemap: fix flaky test (#5581)
The test was failing because the dial attempt would
sometimes timeout and return a different error.

Change the test to first wait for the request to be
received by the server, and then check that context
cancellation is honoured. We do this by explicitly
cancelling the context, rather than using a timeout.