Separate authentication from authorization (#5545)
* Separate authentication from authorization

Overhaul beater/authorization (now beater/auth) to provide
two primary interfaces: Authenticator and Authorizer.

Authenticator is responsible for authenticating a client
(typically but not always an agent), returning
authentication details (e.g. username for API Key) and an
Authorizer for authorizing specific actions.

The auth middleware and interceptor authenticate and store
the returned Authorizer in request context, so it can be
used by specific method handlers. Authorize accepts an
auth.Action and auth.Resource to authorize.

HTTP and gRPC status codes are changed to follow the
semantics more correctly. Failure to authenticate leads to
401 (HTTP) and Unauthenticated (gRPC). Failure to authorize
leads to 403 (HTTP) and PermissionDenied (gRPC).

* tests/system: adjust test for API Key handling

We no longer cache API Key credentials which are
malformatted, i.e. where they're not base64-encoded
id:key. Update the test to match the new behaviour.

* tests/system: adjust to new RUM auth behaviour

We now behave the same for RUM as for backend agents
when auth is not configured. That is, no rate limiting
and no agent config restriction. These only apply when
auth is configured, so update the tests to configure
auth.

* Fix comments