Move more code from model to model processors (#5687)
Move more logic out of model transformation,
into dedicated model processors:

 - modelprocessor.SetExcludeFromGrouping matches
   stack frames against a regular expression,
   setting the "exclude_from_grouping" field

 - modelprocessor.SetLibraryFrame matches stack
   frames against a regular expression, setting the
   "library_frame" field

 - modelprocessor.SetCulprit identifies the first
   source mapped, non-library, stack frame and uses
   it to set or update the error culprit

The "library_frame" field is now only set if it is
true. The UI uses this field but takes omission to
mean the same thing as false (i.e. "falsy")

The new model processors are only installed for
the RUM intake APIs. With these changes, we no
longer need to identify events as originating from
the RUM endpoint, so we remove the model.Span.RUM
and model.Error.RUM fields.