correctly set etag when not defined (#5803)
* correctly set etag when not defined

the calculated etag value wasn't being properly
set on the agent configs. the etag value is handed
to us from kibana so this isn't a big deal, but
when I was testing locally I realized my etag
value was blank, so when I cURL'd the endpoint
without an etag header, they would match (""=="")
and the server would respond with a 304.