model: further align types with document structure (#6066)
Refactor model types to bring them close to the produced
Elasticsearch document structure:

- introduce Parent; remove the ParentID fields from Transaction,
  Span, and Error
- introduce Child; remove the ChildIDs field from from Span
- allow more than one of Transaction, Span, Error, Metricset to be
  set in APMEvent; remove Span and Transaction fields from Metricset;
  remove TransactionID, TransactionSampled, and TransactionType fields
  from  Error; remove TransactionID field from Span

Various transaction and span fields are now only recorded if they
have the non-zero value. We take care to still record zero durations,
as the APM UI makes heavy use of this field and it is not unreasonable
to expect a short span to be recorded as having a zero duration, on
machines which have low resolution timers.