apmbench: Add `eventhandler` and `BenchmarkAgent*` (#7429)
Adds a new `eventhandler.Handler` which allows us to replay stored trace
files to an APM Server using the `embed` package. Loads the traces from
`systemtest/benchtest/events`, and stores them in compressed batches in
memory, these are later replayed when calling `Handler.SendTraces`. It
also has a `WarmUp` function which is called before any of the benchmark
cases are run, allowing the benchmarked APM Server to warm up the local
`sync.Pool` that the APM Server reuses.

Modifies the `apmbench` benchmark executor to:
1. Warm up and wait until the traces have been processed and the APM
   Server is idle.
2. Before starting each benchmark scenario, wait for the server to be
   inactive, or 10 seconds max.
3. Adds `-warmup-events` (default 5000) to control the warmup.
4. Record Transactions, Spans, Metrics and Errors per second from expvar

Last, adds a benchmark scenario the `Go`, `Python`, `NodeJS` and `Ruby`
agents with `BenchmarkAgent<Name>`.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>