processor: Avoid leaking labels between batches (#8081)
Fixes a bug introduced in #6682 by performing a copy of the `baseEvent`
before each of the events is decoded. The event was being shallow copied
and it was reusing the same maps between batches.

The linked PR changed the behavior (although not in a very obviously)
and was instead merging the labels onto the passed `APMEvent`. This
patch reintroduces the previous behavior in a more obvious manner.

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>
Co-authored-by: stuart nelson <stuartnelson3@gmail.com>