beatcmd: make GOMAXPROCS test more robust (#9381)
TestAdjustMaxProcs* tests have been failing
intermittently, particularly on the macOS GitHub
action. The tests assume that the timer will
run at least 10 times, with a ticker duration
of 1ms and total duration of 50ms. This is not
guaranteed, and is causing failures on slow CI
machines.

The two tests are replaced with one test which
checks that duplicate logs are suppressed, and
that changes to the GOMAXPROCS environment variable
are picked up and trigger a log message. The
latter implictly tests that the ticker is effective.
The code-under-test is simple enough that we can
rely on manual code inspection to ensure the
correct duration is used.