agentcfg: implement Elasticsearch fetcher (#9720)
Use Elasticsearch fetcher for agent remote configuration where possible.
Elasticsearch fetcher runs a background goroutine to poll Elasticsearch
to refresh a local agentcfg cache. Agent config fetch requests will be
served from the local cache, from a fallback, or return an error.

The configuration to Elasticsearch is merged from
apm-server.agent.config.elasticsearch over output.elasticsearch over
default ES config.

When apm-server.agent.config.elasticsearch is explicitly configured, no
fallback will ever be used. Otherwise, a fallback will be configured
based on apm-server.kibana.enabled or fleet agent config.

Fetcher uses the fallback when local cache is not ready and a fallback
is configured. The local cache can be not ready due to any reasons
including APM server is starting up and Elasticsearch query failure.

Agent config API returns 403 when local cache is not ready, no fallback
is configured, and Elasticsearch returned a 401 or 403.

The API returns 503 when local cache is not ready, no fallback is
configured, and Elasticsearch did not return a 401 or 403 because it has
not returned yet or it returned another error.