docs: New APM configuration section (#10790)
This PR is mostly structural. It moves the Configure section up two levels in the navigation. This requires many heading-level changes. I also moved all of the configure documentation to a new subdirectory so it's easier to find and edit.

Separating this into its own PR should make the content changes that are coming easier to review.