Minor changes to support running system tests with x-pack-apm (#11638)
* apmpackage: always set {span,transaction}.duration

If the span or transaction field is missing or
null, create a new object rather than omitting
the duration.

* systemtest: changes to work with ES plugin

These are some of the changes needed for the
system tests to pass when using the x-pack-apm
ES plugin templates and pipelines, rather than
the Fleet integration package ones. There are
more changes, which we'll add later when we
remove the data streams from the integration
package.